requirements.txt
numpy==1.26.4
pillow==10.3.0
pygame==2.5.2
